% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay_results_utils.R
\name{archive_assay_results}
\alias{archive_assay_results}
\title{Makes a direct API Call to Benchling Assay Result Archive endpoint}
\usage{
archive_assay_results(
  tenant = Sys.getenv("BENCHLING_TENANT"),
  api_key = Sys.getenv("BENCHLING_API_KEY"),
  assay_result_ids = NULL,
  reason = "Made in error"
)
}
\arguments{
\item{tenant}{tenant name}

\item{api_key}{api key}

\item{assay_result_ids}{is vector type assay result ids. This function supports archiving multiple
items in one function call}

\item{reason}{the reason of the archiving operation. "Made in error" is default but you can change as you wish
based on supported reasons on Benchling}
}
\value{
assay_result_ids
}
\description{
https://benchling.com/api/reference#/Assay%20Results/archiveAssayResults
}
\examples{
\dontrun{
assay_result_ids <- assay_result_ids=c("908df834-220e-4b59-975b-c21bbbb8191a")
reason = "Made in error"
}

}
