% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchling_api_auth.R
\name{benchling_api_auth}
\alias{benchling_api_auth}
\title{Create a Benchling API client that can be used to access the API via R.}
\usage{
benchling_api_auth(tenant, api_key = Sys.getenv("BENCHLING_API_KEY"))
}
\arguments{
\item{tenant}{URL for the Benchling tenant (character).}

\item{api_key}{API key for the Benchling tenant (character). The default
value is the "BENCHLING_API_KEY" environment variable in the `.Renviron`
file.}
}
\value{
Benchling API client object that can be used for accessing the 
API via R.
}
\description{
Create a Benchling API client that can be used to access the API via R.
}
\examples{
\dontrun{
client <- benchling_api_auth(tenant="my-company-tenant-name")
}
}
