% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bl_get_by_id.R
\name{bl_get_by_id}
\alias{bl_get_by_id}
\title{Generic method for getting Benchling entities by identifier.}
\usage{
bl_get_by_id(bl_class, id = NULL, source = "api")
}
\arguments{
\item{bl_class}{Benchling "class".}

\item{id}{Benchling identifier for the object.}

\item{source}{Component of the Benchling developer platform where
the entity should be retrieved from. Entities can be retrieved from
the API (`api`) or data warehouse (`db`).}
}
\description{
Benchling identifiers will typically start with three letters, 
followed by an underscore, and end with a series of alphanumeric 
characters. Custom entities will always start with `bfi_`, while other 
"built-in Benchling types" will start with a specific set of three letters.
Some examples of built-in Benchling types include plates, (`plt`), 
entries (`etr`), and containers (`con`).
}
