% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_sdk_env.R
\name{config_sdk_env}
\alias{config_sdk_env}
\title{Configure environment for accessing API}
\usage{
config_sdk_env(env_name = "benchling-reticulate", env_type = "virtualenv")
}
\arguments{
\item{env_name}{Name for the new environment to be created. The default name
is 'benchling-reticulate'.}

\item{env_type}{'virtualenv' or 'conda'.
(ex. 'options(BENCHLINGR_RETICULATE_ENV = "my_env_name")')}
}
\description{
Helper function to create an Anaconda environment and install
the Benchling Python SDK (`benchling-sdk`) on pip. This Anaconda environment
is used by the reticulate package in R.
}
\examples{
\dontrun{
config_sdk_env()
}
}
