% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_warehouse.R
\name{connect_warehouse}
\alias{connect_warehouse}
\title{Connect to a Benchling data warehouse}
\usage{
connect_warehouse(
  tenant,
  username = Sys.getenv("BENCHLING_WAREHOUSE_USERNAME"),
  password = Sys.getenv("BENCHLING_WAREHOUSE_PASSWORD")
)
}
\arguments{
\item{tenant}{Name of the Benchling tenant for which a warehouse connection
should be opened (character).}

\item{username}{Username for the Benchling tenant (character). The default
argument is the 'BENCHLING_WAREHOUSE_USERNAME' environment variable in the
`.Renviron` file.}

\item{password}{Password for the Benchling tenant (character). The default
argument is the 'BENCHLING_WAREHOUSE_PASSWORD' environment variable in the 
`.Renviron` file.}
}
\value{
A 'database_connection***' object that can be used by the `DBI` package
to make queries against the Postgres database for the specified tenant.
}
\description{
Open a connection to the data warehouse for a particular Benchling tenant
using the `DBI` and `RPostgres` packages.
}
\examples{
\dontrun{
conn <- connect_warehouse("hemoshear-dev", 
    username = Sys.getenv("BENCHLING_WAREHOUSE_USERNAME"),
    password = Sys.getenv("BENCHLING_WAREHOUSE_PASSWORD"))
# It is good practice to close the connection after finishing your queries. 
DBI::dbDisconnect(conn)
}
}
