% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_assay_results.R
\name{create_assay_results}
\alias{create_assay_results}
\title{Upload assay results to Benchling from a data frame}
\usage{
create_assay_results(
  conn,
  client,
  df,
  project_id,
  schema_id,
  fk_type = "name",
  tenant = Sys.getenv("BENCHLING_TENANT"),
  api_key = Sys.getenv("BENCHLING_API_KEY")
)
}
\arguments{
\item{conn}{Database connection.}

\item{client}{Benchling API client.}

\item{df}{Data frame / tibble of results to be uploaded to Benchling.}

\item{project_id}{Benchling project identifier.}

\item{schema_id}{Results schema ID (starts with "assaysch_").}

\item{fk_type}{Are the entity links identifiers or names?}

\item{tenant}{URL for the Benchling tenant.}

\item{api_key}{API key for the Benchling tenant.}
}
\description{
Upload assay results to Benchling from a data frame
}
\examples{
\dontrun{
schema_id <- "assaysch_yKoqVsej"
conn <- warehouse_connect("hemoshear-dev")
api_key <- Sys.getenv("BENCHLING_DEV_API_KEY")
tenant <- "hemoshear-dev"
}
}
