% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disconnect_warehouse.R
\name{disconnect_warehouse}
\alias{disconnect_warehouse}
\title{Disconnect from a Benchling data warehouse}
\usage{
disconnect_warehouse(conn)
}
\arguments{
\item{conn}{A warehouse connection object of class PqConnection usually 
returned by connect_warehouse()}
}
\value{
TRUE, invisibly
}
\description{
Close an already opened connection to the data warehouse.
}
\examples{
\dontrun{
conn <- connect_warehouse(tenant = "hemoshear-dev", 
    username = Sys.getenv("BENCHLING_WAREHOUSE_USERNAME"),
    password = Sys.getenv("BENCHLING_WAREHOUSE_PASSWORD"))
disconnect_warehouse(conn)    
}
}
