% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_entity_by_id.R
\name{.bundle_api_results}
\alias{.bundle_api_results}
\title{Generate a named list of tibble-formatted data tables using the response 
contents from Benchling for specific entity identifiers.}
\usage{
.bundle_api_results(entity_responses, tenant)
}
\arguments{
\item{entity_responses}{A named list where the names are entity identifiers 
and each element is a list that contains the response contents from the call 
made to that identifier's Single-Get API endpoint in Benchling.}

\item{tenant}{A character string containing the name of the Benchling tenant 
for which we want to use to extract information for at least one entity or 
element.}
}
\value{
A named list where the names are entity types and each element is 
another named list where the names are entity schema names and each element 
is a tibble-formatted data table displaying the Single-Get API endpoint 
response contents data extracted Benchling for each entity identifier 
corresponding to a specific entity schema name and entity type.
}
\description{
.bundle_api_results.R takes a named list where the names are entity 
identifiers and each element is a list containing that identifier's 
Single-Get API endpoint response contents from Benchling and sorts the data 
into a named list where the names are the collective entity types for all of 
the entity identifiers in the input and each element is also a named list 
where the names are entity schema names and the elements are tibble-formatted 
data tables containing the Single-Get API endpoint response contents data for 
each entity identifier associated with a specific schema name and entity 
type.
}
\examples{
\dontrun{
entity_id <- c("seq_Cuf0bmCm", "sfs_VUhew7oD", "box_7YutniM0",
"bfi_Ks908uWV", "box_Api455op", "box_t99a7IQ8", "con_m1dmbdV8",
"bfi_ztXInwdh", "bfi_Y7ORWDSz", "bfi_vrXl1zIm", "etr_lu8wTLfL",
"etr_9012RkiU")

inferred_id <- infer_entity_type(entity_id=entity_id, tenant="hemoshear-dev")

entity_single_get_endpoints <- .get_api_endpoints(entity_id=inferred_id,
tenant="hemoshear-dev")

api_queries <- .craft_api_queries(entity_single_get_endpoints=
entity_single_get_endpoints)

entity_responses <- .submit_queries(api_queries=api_queries,
api_key=Sys.getenv("BENCHLING_DEV_API_KEY"))

res <- .bundle_api_results(entity_responses=entity_responses, tenant=
"hemoshear-dev")
}
}
\keyword{internal}
