% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_entity_by_id.R
\name{.craft_api_queries}
\alias{.craft_api_queries}
\title{Format each element using its entity identifier's name to generate proper 
Single-Get API endpoint URLs for that identifier}
\usage{
.craft_api_queries(entity_single_get_endpoints)
}
\arguments{
\item{entity_single_get_endpoints}{A named character vector where the names 
are entity identifiers and the elements are the Single-Get API endpoint URL 
formats for each identifier.}
}
\value{
A named character vector where the names are entity identifiers and 
the elements are the respective Single-Get API endpoint URLs for each 
identifier.
}
\description{
.craft_api_queries.R takes a named character vector with one or more elements 
where the names are entity identifiers and the elements are each identifier's 
respective Single-Get API endpoint URL formats and rewrites the element using 
the name of the identifier to generate proper Single-Get API endpoint URLs.
}
\examples{
\dontrun{
entity_id <- c("seq_Cuf0bmCm", "bfi_Q1PMlXkf", "box_7YutniM0", 
"bfi_Ks908uWV")

inferred_id <- infer_entity_type(entity_id=entity_id, tenant="hemoshear-dev")

entity_single_get_endpoints <- .get_api_endpoints(entity_id=inferred_id,
tenant="hemoshear-dev")

res <- .craft_api_queries(entity_single_get_endpoints=
entity_single_get_endpoints)
}
}
\keyword{internal}
