% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_results.R
\name{.create_assay_result}
\alias{.create_assay_result}
\title{Create an assay result to be uploaded to Benchling}
\usage{
.create_assay_result(
  project_id,
  schema_id,
  fields,
  id = NULL,
  field_validation = NULL
)
}
\arguments{
\item{project_id}{Identifier for the Benchling project}

\item{fields}{List where the keys are strings corresponding to the field names. The values
should be lists with at least one element named `value`, which is the value that should be submitted
for that field. 
ex. list(study = list(value = "MyFakeStudy"))}

\item{id}{UUID for result. Optional}

\item{field_validation}{Field validation for result. Optional.}

\item{schema}{Identifier for the result schema}
}
\value{
List to be passed to `upload_assay_result`.
}
\description{
Create an assay result to be uploaded to Benchling
}
\keyword{internal}
