% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_entity_by_id.R
\name{.get_api_endpoints}
\alias{.get_api_endpoints}
\title{Match each element to its corresponding Single-Get API endpoint URL format}
\usage{
.get_api_endpoints(entity_id, tenant)
}
\arguments{
\item{entity_id}{A named character vector where the names are entity 
identifiers and the elements are the entity types associated with those 
identifiers.}

\item{tenant}{A character string containing the name of the Benchling tenant 
for which we want to use to extract information for at least one entity or 
element.}
}
\value{
A named character vector where the names are entity identifiers and 
the elements are the Single-Get API endpoint URL formats associated with them 
based on the respective entities each identifier correlates to and matches 
with.
}
\description{
.get_api_endpoints.R takes a named character vector with one or more elements 
where the names are entity identifiers and the elements are the entity types 
for those identifiers and matches each element to their respective Single-Get 
API endpoint URL formats according to which entity the schema correlates 
with.
}
\examples{
\dontrun{
entity_id1 <- c("seq_Cuf0bmCm", "bfi_Q1PMlXkf", "box_7YutniM0")
inferred_id1 <- infer_entity_type(entity_id=entity_id1, tenant=
"hemoshear-dev")
res1 <- .get_api_endpoints(entity_id=inferred_id1, tenant="hemoshear-dev")

entity_id2 <- c("bfi_Ur5DfvGJ", "seq_Gju61mCm", "bfi_Q13AlXkf", 
"bfi_Ks908uWV", "ent_Ec76qX9f", "ent_sPrxBNOh", "box_K9950IQ8")
inferred_id2 <- infer_entity_type(entity_id=entity_id2, tenant=
"hemoshear-dev")
res2 <- .get_api_endpoints(entity_id=inferred_id2, tenant="hemoshear-dev")
}
}
\keyword{internal}
