% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_entity_table.R
\name{.get_entity_table}
\alias{.get_entity_table}
\title{Get additional information about entities in a warehouse table (single entity column.)}
\usage{
.get_entity_table(conn, df, column, return_cols = c("id", "name$"))
}
\arguments{
\item{conn}{Database connection opened by `warehouse_connect`}

\item{df}{Data frame with one or more entity columns. The data frame
must also have a column called `schema`, which indicates the schema
name of the warehouse table.}

\item{column}{Name of the entity column.}

\item{return_cols}{Character vector of warehouse column names to include
in results. If `return_cols` is missing, then every column will be returned.}
}
\value{
data frame with rows from warehouse table that correspond to 
entities found in the input data frame (`df`).
}
\description{
Given a data frame retrieved from the Benchling warehouse, extract the
additional table in the warehouse that correspond to the entities in the 
data frame.
}
\keyword{internal}
