% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_entity_by_id.R
\name{.handle_api_responses}
\alias{.handle_api_responses}
\title{Check each element's API response and remove any elements that have errors in
their responses.}
\usage{
.handle_api_responses(entity_id_responses)
}
\arguments{
\item{entity_id_responses}{A named list where the names are entity 
identifiers and the elements are each entity identifier's Single-Get API 
endpoint response contents from Benchling.}
}
\value{
A named list where the names are valid entity identifiers that do not 
possess errors in their Single-Get API endpoint response contents and the
elements are said identifier's Single-Get API endpoint response contents.
}
\description{
.handle_api_response.R takes a named list where the names are entity 
identifiers and the elements are each identifier's Single-Get API endpoint 
response contents extracted from Benchling and analyzes it to see if any 
entity identifiers contain errors in their Single-Get API endpoint responses 
and removes them from the list.
}
\examples{
\dontrun{
entity_id <- c("seq_Cuf0bmCm", "bfi_Q1PMlXkf", "box_7YutniM0", 
"bfi_Ks908uWV", "box_Api455op", "box_t99a7IQ8", "con_m1dmbdV8", 
"bfi_ztXInwdh", "bfi_Y7ORWDSz", "bfi_vrXl1zIm", "etr_lu8wTLfL", 
"etr_9012RkiU")

inferred_id <- infer_entity_type(entity_id=entity_id, tenant="hemoshear-dev")

entity_single_get_endpoints <- .get_api_endpoints(entity_id=inferred_id,
tenant="hemoshear-dev")

api_queries <- .craft_api_queries(entity_single_get_endpoints=
entity_single_get_endpoints)

entity_id_responses <- .submit_queries(api_queries=api_queries, 
api_key=Sys.getenv("BENCHLING_API_KEY"))

res <- .handle_api_responses(entity_id_responses=entity_id_responses)
}
}
\keyword{internal}
