% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_entity_type.R
\name{.infer_entity_type}
\alias{.infer_entity_type}
\title{Infer the entity types of elements contained in a vector or list}
\usage{
.infer_entity_type(entity_id, verbose = F)
}
\arguments{
\item{entity_id}{Either a character vector or list that contains the entities. 
All entities in the list or character vector must be character strings.}

\item{verbose}{Messages}
}
\value{
A named list, where the names are the identifiers or elements from entity_id
and the values are the schema types, single API endpoint and bulk-get API endpoints.
}
\description{
infer_entity_type.R takes either a character vector or list of character strings and 
for each element tries to infer the schema type, single-get API Endpoint URL format and bulk-get 
API endpoints URL format using string-matching techniques.
}
\examples{
\dontrun{
entity_id <- c("seq_Cuf0bmCm", "bfi_Q1PMlXkf")
res <- infer_entity_type(entity_id)
}
}
\keyword{internal}
