% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_api_contents.R
\name{.list_api_contents}
\alias{.list_api_contents}
\title{Generate a named list where the names are the first characters in an 
identifier's name for a specific entity and each element is a vector with the 
first element in the vector being the entity schema, the second vector is the 
list API contents, the third is the Single-Get API endpoint option and the 
fourth is the Bulk-Get API endpoints option.}
\usage{
.list_api_contents(tenant, contents = "all")
}
\arguments{
\item{tenant}{A character string containing the name of the Benchling tenant 
for which we want to use to extract information for at least one entity or 
element.}

\item{contents}{A character string to specify which type of information to 
return for the entities listed: `"bulk-get endpoints"`, 
`"single-get endpoint"`,  `"list contents"`, `"entity schema"`, 
`"entity type"`, `"all"` (the default).}
}
\value{
A named list where the names are the first characters in an 
identifier's name for a corresponding entity and the elements are vectors 
that can show either the entity's schema, list API contents, Single-Get API 
endpoint option, Bulk-Get API endpoints option or all of them respectively.
}
\description{
Generate a named list where the names are the first characters in an 
identifier's name for a specific entity and each element is a vector with the 
first element in the vector being the entity schema, the second vector is the 
list API contents, the third is the Single-Get API endpoint option and the 
fourth is the Bulk-Get API endpoints option.
}
\examples{
\dontrun{
entity_list <- .list_api_contents(tenant="hemoshear-dev")

entity_schemas <- .list_api_contents(tenant="hemoshear-dev", 
contents="entity schema")

entity_list_contents <- .list_api_contents(tenant="hemoshear-dev", 
contents="list contents")

entity_single_get_contents <- .list_api_contents(tenant="hemoshear-dev", 
contents="single-get endpoint")

entity_bulk_get_contents <- .list_api_contents(tenant="hemoshear-dev", 
contents="bulk-get endpoints")
}
}
\keyword{internal}
