% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_entity_field_names_to_warehouse_tables.R
\name{.map_entity_field_names_to_warehouse_tables}
\alias{.map_entity_field_names_to_warehouse_tables}
\title{Map entity field names in one warehouse table to corresponding warehouse
table with the relevant entities.}
\usage{
.map_entity_field_names_to_warehouse_tables(conn, df)
}
\arguments{
\item{conn}{Database connection opened by `connect_warehouse`}

\item{df}{data.frame with one or more fields that are entity types.}
}
\value{
A vector where the names are the warehouse names for the fields 
in the schema and the values are the warehouse names for the entities in the 
schema
}
\description{
A schema field name can differ from the schema name itself, so this function
uses the `schema` and `schema_field` tables in the warehouse to link 
the field name in one schema to the actual name of the entity.
}
\examples{
\dontrun{
conn <- connect_warehouse(
"hemoshear-dev",
 username = Sys.getenv("BENCHLING_DEV_WAREHOUSE_USERNAME"),
 password = Sys.getenv("BENCHLING_DEV_WAREHOUSE_PASSWORD"))
 
 d <- DBI::dbGetQuery(conn, "SELECT * FROM simple_plate_analyte_mapping$raw")
 .map_entity_field_names_to_warehouse_tables(conn, d)
 }
}
\keyword{internal}
