% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.R
\name{.missing_api_key_error}
\alias{.missing_api_key_error}
\title{Stop function execution if API key is an empty string.}
\usage{
.missing_api_key_error()
}
\value{
Nothing.
}
\description{
Stop function execution if API key is an empty string.
}
\keyword{internal}
