% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.R
\name{.missing_tenant_error}
\alias{.missing_tenant_error}
\title{Stop function execution if tenant is an empty string.}
\usage{
.missing_tenant_error()
}
\value{
Nothing.
}
\description{
Stop function execution if tenant is an empty string.
}
\keyword{internal}
