% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_plate_diagrams.R
\name{.parse_plate_diagram}
\alias{.parse_plate_diagram}
\title{Parse well layout JSON from a notebook entry}
\usage{
.parse_plate_diagram(json)
}
\arguments{
\item{json}{Well layout JSON.}
}
\value{
Long form table with the following columns: `row`, `column`, `content`, and `plate`.
}
\description{
Parse well layout JSON from a notebook entry
}
