% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_entity_by_id.R
\name{.submit_queries}
\alias{.submit_queries}
\title{Uses each element to call the Single-Get API endpoint in Benchling for a 
specific entity identifier.}
\usage{
.submit_queries(api_queries, api_key = Sys.getenv("BENCHLING_API_KEY"))
}
\arguments{
\item{api_queries}{A named character vector where the names are entity 
identifiers and the elements are the Single-Get API endpoint URLs for each 
identifier.}

\item{api_key}{Benchling API key.}
}
\value{
A named list where the names are entity identifiers and each element 
is a list that contains the response contents from the call made to that 
identifier's Single-Get API endpoint in Benchling.
}
\description{
.submit_queries.R takes a named character vector with one or more elements 
where the names are entity identifiers and the elements are each identifier's 
respective Single-Get API endpoint URL and uses them to call that 
identifier's Single-Get API endpoint in Benchling and extract the response 
contents.
}
\examples{
\dontrun{
entity_id <- c("seq_Cuf0bmCm", "bfi_Q1PMlXkf", "box_7YutniM0", 
"bfi_Ks908uWV", "box_Api455op", "box_t99a7IQ8", "con_m1dmbdV8", 
"bfi_ztXInwdh", "bfi_Y7ORWDSz", "bfi_vrXl1zIm", "etr_lu8wTLfL", 
"etr_9012RkiU")

inferred_id <- infer_entity_type(entity_id=entity_id, tenant="hemoshear-dev")

entity_single_get_endpoints <- .get_api_endpoints(entity_id=inferred_id,
tenant="hemoshear-dev")

api_queries <- .craft_api_queries(entity_single_get_endpoints=
entity_single_get_endpoints)

res <- .submit_queries(api_queries=api_queries, 
api_key=Sys.getenv("BENCHLING_API_KEY"))
}
}
\keyword{internal}
