% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_files.R
\name{.upload_files}
\alias{.upload_files}
\title{Upload a character vector}
\usage{
.upload_files(file, client, blob_link_cols = NULL)
}
\arguments{
\item{file}{Character vector, list, or data frame with file names
to be uploaded.}

\item{client}{Benchling Python SDK object}

\item{blob_link_cols}{If a data frame is provided, the names of
the blob link columns must be specified.}
}
\description{
Upload a character vector
}
\keyword{internal}
