% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_validation.R
\name{.validate_blob_link_column_values}
\alias{.validate_blob_link_column_values}
\title{Verify that values in a blob link column refer to valid file paths on the 
local machine or that identifiers already exist.}
\usage{
.validate_blob_link_column_values(
  client,
  errors,
  values,
  column_name,
  multi_select,
  fk_type
)
}
\arguments{
\item{errors}{Errors}

\item{values}{Values in the column.}

\item{column_name}{Name of the column in the input data frame.}

\item{fk_type}{"id" or "name". Use "id" if the blob already exists on 
Benchling and you are submitting the ID for the blob. Use "name" if
the file doesn't already exist on Benchling as a blob and you need to upload
the file to Benchling.}
}
\description{
Verify that values in a blob link column refer to valid file paths on the 
local machine or that identifiers already exist.
}
\keyword{internal}
