% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_results.R
\name{.validate_column_types}
\alias{.validate_column_types}
\title{Ensure that column types are valid for a data frame that is to be uploaded
to Benchling as a results table.}
\usage{
.validate_column_types(errors, values, column_name, type_map, multi_select_map)
}
\description{
integer|float --> numeric
text|entity_link|dropdown|long_text|storage_link|blob_link|dna_sequence_link --> character
}
\keyword{internal}
