% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_validation.R
\name{.validate_column_values}
\alias{.validate_column_values}
\title{Validate the values in the column of a results table.}
\usage{
.validate_column_values(
  conn,
  client,
  errors,
  values,
  column_name,
  benchling_type,
  multi_select,
  fk_type,
  target_schema_id
)
}
\arguments{
\item{conn}{Database connection typically opened with `warehouse_connect`.}

\item{client}{A Benchling SDK object typically created with `benchling_api_auth`.}

\item{errors}{Character vector of errors.}

\item{values}{Values in the column.}

\item{column_name}{Name of the column in the data frame to be uploaded to Benchling.}

\item{benchling_type}{The Benchling "type" that the column corresponds to.}

\item{multi_select}{Boolean indicating whether or not the column corresponds
to a multi-select field.}

\item{fk_type}{String "id" or "name$" which indicates whether the column
represents the Benchling ID or name of the entities.}

\item{target_schema_id}{The name of the warehouse table which corresponds
to the column.}
}
\description{
If the column refers to an entity link field, the entity IDs in the input
data frame are checked against those currently in the registry. If the column
is a dropdown, then the values are checked against the dropdown options. 
If the column is a storage link, the function checks to see if the IDs in the 
input data frame exist in the inventory. If the column refers to a blob
link field, then the function will assume the values are file paths and will
check to see if the file exists on the local machine.
}
\keyword{internal}
