% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_validation.R
\name{.validate_data_frame}
\alias{.validate_data_frame}
\title{Raise exception if the `schema` column is not in the data frame.
 Dataframe validation}
\usage{
.validate_data_frame(conn, client, df, fk_type = "name", mappings)
}
\arguments{
\item{conn}{Data warehouse connection typically opened by `connect_warehouse`.}

\item{client}{API facade object}

\item{df}{data.frame with table from the data warehouse.}
}
\value{
error list
}
\description{
Raise exception if the `schema` column is not in the data frame.
 Dataframe validation
}
\keyword{internal}
