% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_validation.R
\name{.validate_dropdown_column_values}
\alias{.validate_dropdown_column_values}
\title{Verify that values in dropdown column are valid options in the dropdown schema.}
\usage{
.validate_dropdown_column_values(
  conn,
  errors,
  values,
  column_name,
  dropdown_id
)
}
\description{
Verify that values in dropdown column are valid options in the dropdown schema.
}
\keyword{internal}
