% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec2sql_tuple.R
\name{.vec2sql_tuple}
\alias{.vec2sql_tuple}
\title{Convert a vector into a valid tuple that can be used
in SQL queries.}
\usage{
.vec2sql_tuple(x)
}
\arguments{
\item{x}{Character or numeric vector}
}
\value{
A character vector that represents a SQL tuple to be
used in SQL queries.
}
\description{
Convert a vector into a valid tuple that can be used
in SQL queries.
}
\examples{
x <- c('A', 'B', 'C')
res <- .vec2sql_tuple(x)
y <- c(1, 2)
res <- .vec2sql_tuple(y)
}
