% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_files.R
\name{download_blobs}
\alias{download_blobs}
\title{Download file attachments from Benchling}
\usage{
download_blobs(client, file_map, outdir, condaenv = "benchling-reticulate")
}
\arguments{
\item{client}{Benchling client object created by `benchling_api_auth()`.}

\item{file_map}{List where names are blob (file) identifiers
and values are file names.}

\item{outdir}{Directory in which to save output files.}

\item{condaenv}{Name of the conda environment to use. Defaults to 
`benchling-reticulate`.}
}
\description{
Download "blobs" attached to Benchling entities or results.
}
\examples{
\dontrun{
conn <- warehouse_connect("hemoshear-dev", 
    username = Sys.getenv("BENCHLING_DEV_WAREHOUSE_USERNAME"),
    password = Sys.getenv("BENCHLING_DEV_WAREHOUSE_PASSWORD"))
res <- DBI::dbGetQuery(conn, "SELECT * FROM 
non_barcode_plate_file_results$raw WHERE study_codes::text LIKE 
'\%bfi_jUgN7fPL\%'")
file_col <- purrr::map(as.character(res$excel_file),
                      ~ RJSONIO::fromJSON(.) \%>\% .[[1]])
file_map <- purrr::map_chr(file_col, ~ .['name'])
names(file_map) <- purrr::map_chr(file_col, ~ .['id'])
file_map <- as.list(file_map)
download_blobs(file_map, outdir='data')
}

}
