% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_files.R
\name{download_blobs_in_warehouse_table}
\alias{download_blobs_in_warehouse_table}
\title{Download blobs contained within a warehouse table.}
\usage{
download_blobs_in_warehouse_table(
  client,
  conn,
  df,
  columns = NULL,
  outdir = NULL,
  outdir_column = NULL
)
}
\arguments{
\item{client}{A Benchling API client object.}

\item{conn}{A warehouse connection opened with `warehouse_connect`.}

\item{df}{Data frame retrieved from the Benchling data warehouse.
Must have the `schema` column included.}

\item{columns}{Character vector of column names in the table to download 
blobs from.}

\item{outdir}{Directory where the files should be saved on the local machine.
Subdirectories are created within this directory for each blob link column
in the data frame. If `outdir_column` is provided, then this argument is 
ignored.}

\item{outdir_column}{Column in the data frame that represents the directories
in which the files should be saved. This argument overrides `outdir`.}
}
\description{
Download blobs contained within a warehouse table.
}
\examples{
\dontrun{
library(magrittr)
client <- benchling_api_auth("https://hemoshear.benchling.com")
conn <- warehouse_connect("hemoshear-dev", 
    username = Sys.getenv("BENCHLING_DEV_WAREHOUSE_USERNAME"),
    password = Sys.getenv("BENCHLING_DEV_WAREHOUSE_PASSWORD"))
res <- DBI::dbGetQuery(conn, "SELECT * FROM 
 non_barcode_plate_file_results$raw WHERE study_codes::text LIKE 
 '\%bfi_jUgN7fPL\%'")
download_blobs_in_warehouse_table(conn, res, outdir='temp_data_dir/')
}
}
