% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_files.R
\name{download_files}
\alias{download_files}
\title{Download file attachments from Benchling}
\usage{
download_files(
  file_map,
  outdir,
  api_key = Sys.getenv("BENCHLING_API_KEY"),
  condaenv = "benchling-reticulate"
)
}
\arguments{
\item{file_map}{List where names are blob (file) identifiers
and values are file names.}

\item{outdir}{Directory in which to save output files.}

\item{api_key}{Benchling API key. Defaults to the `BENCHLING_API_KEY` 
environment variable in `.Renviron`.}

\item{condaenv}{Name of the conda environment to use. Defaults to 
`benchling-reticulate`.}
}
\description{
Download file attachments from Benchling
}
\examples{
\dontrun{
res <- DBI::dbGetQuery(conn, "SELECT * FROM non_barcode_plate_file_results$raw WHERE study_codes::text LIKE '\%bfi_jUgN7fPL\%'")
file_col <- purrr::map(as.character(res$excel_file),
                      ~ RJSONIO::fromJSON(.) \%>\% .[[1]])
file_map <- purrr::map_chr(file_col, ~ .['name'])
names(file_map) <- purrr::map_chr(file_col, ~ .['id'])
file_map <- as.list(file_map)
download_files(file_map, outdir='data')
}

}
