% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay_results_utils.R
\name{get_assay_result}
\alias{get_assay_result}
\title{Makes a direct API Call to Benchling Get Assay Result endpoint.}
\usage{
get_assay_result(
  tenant = Sys.getenv("BENCHLING_TENANT"),
  api_key = Sys.getenv("BENCHLING_API_KEY"),
  assay_result_id = ""
)
}
\arguments{
\item{tenant}{tenant name}

\item{api_key}{api key}

\item{assay_result_id}{is the id of the assay result.}
}
\value{
assay result details
}
\description{
https://benchling.com/api/reference#/Assay%20Results/getAssayResult
}
\examples{
\dontrun{
assay_result_id <- '908df834-220e-4b59-975b-c21bbbb8191a'
}
}
