% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_assay_results.R
\name{get_assay_results_schema_ids}
\alias{get_assay_results_schema_ids}
\title{Get Benchling assay results metadata}
\usage{
get_assay_results_schema_ids(conn)
}
\arguments{
\item{conn}{Database connection opened with `connect_warehouse`.}
}
\value{
data.frame with `id` and `name` attributes for Benchling assay results
}
\description{
Get Benchling assay results metadata
}
