% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropdown.R
\name{get_dropdown}
\alias{get_dropdown}
\title{Get metadata for a dropdown menu}
\usage{
get_dropdown(conn, name)
}
\arguments{
\item{conn}{Database connection opened with `connect_warehouse`.}

\item{name}{Schema name for the dropdown}
}
\description{
Get metadata for a dropdown menu
}
