% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_entity_by_id.R
\name{get_entity_by_id}
\alias{get_entity_by_id}
\title{Calls each element's Single-Get API endpoints in Benchling and extracts the 
response.}
\usage{
get_entity_by_id(
  entity_id,
  api_key = Sys.getenv("BENCHLING_API_KEY"),
  tenant,
  bundle = TRUE
)
}
\arguments{
\item{entity_id}{A character vector with 1 or more elements.}

\item{api_key}{Benchling API key.}

\item{tenant}{A character string containing the name of the Benchling tenant 
for which we want to use to extract information for at least one entity or 
element.}

\item{bundle}{A boolean that determines how the output is organized.
If `TRUE` is selected, the output will be organized into a named list where 
the names are entity types and each element is a named list of 
tibble-formatted data tables where the names are schema names and each data
table displays the API response contents output extracted from Benchling for 
each entity identifier inside. If `FALSE`, then the output will be made into 
a named list where the names are entity identifiers and each element contains
the API response contents from the call made to that identifier's Single-Get 
API endpoint in Benchling.}
}
\value{
Either a named list where the names are entity identifiers and the 
elements are each identifier's API response contents extracted from 
Benchling or a named list where the names are entity types and each element 
is a named list of tibble-formatted tables where the names are schema names
and each table features the API response contents for the identifiers 
specified under both the entity type and schema name.
}
\description{
get_entity_by_id.R takes a character vector with one or more elements and for 
each element calls its Single-Get API endpoints in Benchling and extracts the 
response contents and stores it in a list.
}
\examples{
\dontrun{
entity_id <- c("bfi_Ur5DfvGJ", "seq_Gju61mCm", "bfi_Q13AlXkf", 
"bfi_Ks908uWV", "ent_Ec76qX9f", "ent_sPrxBNOh", "box_K9950IQ8", 
"box_923aklum", "ver_asd89230", "ysq_983mnK4e")
res <- get_entity_by_id(entity_id=entity_id, 
api_key=Sys.getenv("BENCHLING_API_KEY"), tenant="hemoshear-dev",
bundle=TRUE)
}
}
