% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_entity_name_from_id.R
\name{get_entity_name_from_id}
\alias{get_entity_name_from_id}
\title{Retrieves the name attribute from each element}
\usage{
get_entity_name_from_id(entity_response_list)
}
\arguments{
\item{entity_response_list}{A named list where the names are entity identifiers and each element is a list
containing the response contents from the call made to that identifier's Single-Get API endpoint in Benchling.}
}
\value{
A named character vector where the names are entity identifiers and the elements are the names
corresponding to those identifiers.
}
\description{
get_entity_name_from_id.R takes a named list where the names are entity identifiers and each element in the list
is a list as well that contains the contents of the response from the call to each identifier's Single-Get API endpoint 
in Benchling and extracts the name associated with it.
}
\examples{
\dontrun{
entity_id <- c("seq_Cuf0bmCm", "bfi_Ur5DfvGJ", "seq_Gju61mCm", "bfi_Q13AlXkf", "bfi_Ks908uWV", 
"ent_Ec76qX9f", "ent_sPrxBNOh", "box_K9950IQ8", "box_923aklum")
entity_responses <- get_entity(entity_id=entity_id, api_key=Sys.getenv("BENCHLING_API_KEY"))

res <- get_entity_name_from_id(entity_response_list=entity_responses)
}
}
