% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_entry_table.R
\name{get_entry_table}
\alias{get_entry_table}
\title{Read unstructured table in a notebook entry}
\usage{
get_entry_table(
  entry,
  day = NULL,
  table_position = NULL,
  return_table_name = TRUE
)
}
\arguments{
\item{entry}{Notebook entry in JSON format.}

\item{day}{Integer for the day in the notebook entry. See `list_entry_tables`.}

\item{table_position}{Integer for the position of the table in the notebook entry list. 
See `list_entry_tables`.}

\item{return_table_name}{bool Determines how the name of the table in the notebook 
entry is included in the output.
If `TRUE` is selected, the table name will be added as a new column
to the data frame. If `FALSE`, then the table name will be ignored.}
}
\value{
Data frame representing the unstructured table in the notebook
entry.
}
\description{
Read unstructured table in a notebook entry
}
\examples{
\dontrun{
client <- benchlingr::connect_sdk(
    tenant="https://hemoshear-dev.benchling.com",
    api_key = Sys.getenv("BENCHLING_DEV_API_KEY"))
entry <- client$entries$get_entry_by_id("etr_T3WZTyAe")

table_indices <- benchlingr:::list_entry_tables(entry)
print(table_indices)
a_table <- get_entry_table(entry, day=1, table_position=2)
}
}
\keyword{internal}
