% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plate_diagrams.R
\name{get_plate_diagrams}
\alias{get_plate_diagrams}
\title{Find and parse plate diagrams in a notebook entry}
\usage{
get_plate_diagrams(entry, plate_dim = NULL)
}
\arguments{
\item{entry}{response from GET /entry/{entry_id}}

\item{plate_dim}{A numeric vector of length 2 that describes the dimensions
of the plates that should be extracted from the notebook entry. For example,
`c(16, 24)` will return only 384-well plates. The plate sizes
must be `c(2,3)`, `c(3,4)`, `c(4,6)`, `c(6,8)`, `c(8, 12)`, `c(16, 24)`, 
or `c(32, 48)`. By default, `plate_dim` is `NULL` the function will
return plates of all sizes supported by Benchling.}
}
\value{
List of data frames representing the plate diagrams present
in the notebook entry.
}
\description{
Benchling users can add special unstructured tables to notebook entrys
called "plate diagrams" for conveying information about the contents
of physical plates. The `read_plate_layouts` function will extract these 
plate diagrams into data frames for further processing in R.
}
\examples{
\dontrun{
client <- connect_sdk(tenant="https://hemoshear-dev.benchling.com",
                            api_key=Sys.getenv("BENCHLING_DEV_API_KEY"))
entry <- client$entries$get_entry_by_id("etr_f1bpDIes")
plate_diagrams <- get_plate_diagrams(entry)
}
}
