% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_results.R
\name{get_project_ids}
\alias{get_project_ids}
\title{Get Benchling project metadata}
\usage{
get_project_ids(conn)
}
\arguments{
\item{conn}{Database connection opened with `warehouse_connect`.}
}
\value{
data.frame with `id` and `name` attributes for Benchling projects.
}
\description{
Get Benchling project metadata
}
