% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema_utils.R
\name{get_schema_fields}
\alias{get_schema_fields}
\title{Makes a direct API Call to Benchling Schema endpoints without benchling-sdk client.}
\usage{
get_schema_fields(schema_id, schema_type, tenant)
}
\arguments{
\item{schema_id}{provided schema id}

\item{schema_type}{schema type name as it shown on benhcling API endpoints with or without `-schemas` ending.}

\item{tenant}{is tenant name. If missing, it will be reading from sys.env}
}
\description{
Makes a direct API Call to Benchling Schema endpoints without benchling-sdk client.
}
\examples{
\dontrun{
schema_id <- "assaysch_nIw4yAq8"
schema_type <- "assay-result"
}

}
