% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{is_schema_in_dataframe}
\alias{is_schema_in_dataframe}
\title{Raise exception if the `schema` column is not in the data frame.}
\usage{
is_schema_in_dataframe(df)
}
\arguments{
\item{df}{data.frame with table from the data warehouse.}
}
\value{
Used for side effect only.
}
\description{
The `schema` column appears in every data warehouse table. This function
stops function execution if `schema` is not in the data frame.
}
