% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_entry_tables.R
\name{list_entry_tables}
\alias{list_entry_tables}
\title{Find unstructured tables (not registry or results) in a notebook entry.}
\usage{
list_entry_tables(entry, min_rows = NULL)
}
\arguments{
\item{entry}{entry response from GET /entry/{entry_id}. *** Link the appropriate
Benchling API endpoint here.}

\item{min_rows}{Keep tables with this number of rows or more.}
}
\value{
List where names are days and elements are indices in the `notes` attribute that
   correspond to tables in the entry.
}
\description{
This function returns the locations of unstructured tables in a notebook entry.
It does not show the user where any registration or results tables might be 
located in a notebook entry, as information in those tables must be retrieved
from other API endpoints / data warehouse tables.
}
\examples{
\dontrun{
client <- benchlingr::connect_sdk(
    tenant="https://hemoshear-dev.benchling.com",
    api_key = Sys.getenv("BENCHLING_DEV_API_KEY"))
entry <- client$entries$get_entry_by_id("etr_T3WZTyAe")
benchlingr::list_entry_tables(entry)
}
}
