% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_multiselect_column.R
\name{list_multiselect_columns}
\alias{list_multiselect_columns}
\title{List the multi-select columns in a data frame retrieved from the data 
warehouse}
\usage{
list_multiselect_columns(conn, df)
}
\arguments{
\item{conn}{Database connection opened with `warehouse_connect`.}

\item{df}{Data frame retrieved from the data warehouse.}
}
\value{
Not sure yet
}
\description{
List the multi-select columns in a data frame retrieved from the data 
warehouse
}
