% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_schemas_in_entry.R
\name{list_schemas_in_entry}
\alias{list_schemas_in_entry}
\title{List the schemas for the results and registration tables contained within
a notebook entry}
\usage{
list_schemas_in_entry(client, conn, entry)
}
\arguments{
\item{client}{Benchling API client created by `benchling_api_auth`.}

\item{conn}{Benchling data warehouse connection created with `warehouse_connect`.}

\item{entry}{Notebook entry retrieved with `client$entries$get_entry_by_id`.}
}
\value{
A data frame that includes the names and identifiers of the results
and registration schemas contained within a notebook entry.
}
\description{
List the schemas for the results and registration tables contained within
a notebook entry
}
\examples{
\dontrun{
conn <- warehouse_connect("hemoshear-dev", 
                         username = Sys.getenv("BENCHLING_DEV_WAREHOUSE_USERNAME"),
                         password = Sys.getenv("BENCHLING_DEV_WAREHOUSE_PASSWORD"))
client <- benchling_api_auth(tenant="https://hemoshear-dev.benchling.com",
                            api_key=Sys.getenv("BENCHLING_DEV_API_KEY"))
nb_entry <- client$entries$get_entry_by_id("etr_MWQ7M7Pz")
list_schemas_in_entry(client, conn, nb_entry)
}
}
