% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_entry_tables.R
\name{read_entry_table}
\alias{read_entry_table}
\title{Read unstructured table in a notebook entry}
\usage{
read_entry_table(entry, day, table_position, table_name = TRUE)
}
\arguments{
\item{entry}{Notebook entry in JSON format.}

\item{day}{Integer for the day in the notebook entry. See `find_entry_tables`.}

\item{table_position}{Integer for the position of the table in the notebook entry list. 
See `find_entry_tables`.}

\item{table_name}{bool Determines how
the name of the table in the notebook entry is included in the output.
If `TRUE` is selected, the table name will be added as a new column
to the data frame. If `FALSE`, then the table name will be ignored.}
}
\value{
Data frame representing the unstructured table in the notebook
entry.
}
\description{
Read unstructured table in a notebook entry
}
\examples{
\dontrun{
client <- benchling_api_auth(tenant="https://hemoshear-dev.benchling.com",
                             api_key=Sys.getenv("BENCHLING_DEV_API_KEY"))
entry <- client$entries$get_entry_by_id("etr_T3WZTyAe")
table_indices <- benchlingr:::find_entry_tables(entry)
print(table_indices)
a_table <- read_entry_table(entry, day=1, table_position=2)
}
}
\keyword{internal}
