% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_plate_diagrams.R
\name{read_plate_diagrams}
\alias{read_plate_diagrams}
\title{Find and parse plate diagrams in a notebook entry}
\usage{
read_plate_diagrams(json, min_rows = NULL, flatten = TRUE, verbose = TRUE)
}
\arguments{
\item{json}{response from GET /entry/{entry_id}}

\item{min_rows}{Minimum number of rows on each plate.}

\item{flatten}{If flatten is TRUE, then the results are collapsed into a 
single data frame, in which the `entry_index` column shows which day the 
plate was read from. If flatten is FALSE, then a nested list is returned, 
where each element corresponds to a single day.}

\item{verbose}{If verbose, then the table names are printed to the console.}
}
\value{
List of data frames representing the plate diagrams present
in the notebook entry.
}
\description{
Benchling users can add special unstructured tables to notebook entrys
called "plate diagrams" for conveying information about the contents
of physical plates. The `read_plate_layouts` function will extract these 
plate diagrams into data frames for further processing in R.
}
