% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_files.R
\name{upload_files.data.frame}
\alias{upload_files.data.frame}
\title{Upload files represented by one or more columns in a data frame containing
file paths.}
\usage{
\method{upload_files}{data.frame}(file, client, blob_link_cols = NULL)
}
\arguments{
\item{file}{data.frame with one or more columns containing files 
to be uploaded.}

\item{blob_link_cols}{Character vector of column names containing files
to be uploaded}
}
\description{
Upload files represented by one or more columns in a data frame containing
file paths.
}
\keyword{internal}
