% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_results.R
\name{upload_results}
\alias{upload_results}
\title{Upload assay results to Benchling from a data frame}
\usage{
upload_results(
  df,
  project_id,
  schema_id,
  tenant = Sys.getenv("BENCHLING_TENANT"),
  id_or_name = "name"
)
}
\arguments{
\item{df}{Data frame / tibble of results to be uploaded to Benchling.}

\item{conn}{Database connection.}

\item{client}{Benchling API client.}

\item{projectId}{Benchling project identifier.}

\item{schemaId}{Results schema ID (starts with "assaysch_").}
}
\description{
Upload assay results to Benchling from a data frame
}
\examples{
\dontrun{
schemaId <- "assaysch_BoT5QoQc"
}

}
