% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema_utils.R
\name{verify_schema_fields}
\alias{verify_schema_fields}
\title{compares schema fields with dataframe colnames and verifies}
\usage{
verify_schema_fields(schema_id, schema_type, df, strict_check = FALSE, tenant)
}
\arguments{
\item{schema_id}{provided schema id}

\item{schema_type}{schema type name as it shown on benhcling API endpoints with or without `-schemas` ending.}

\item{df}{provided dataframe}

\item{strict_check}{is an optional arguement. if set TRUE, the function looks for every schema field names in colnames.}

\item{tenant}{is tenant name. If missing, it will be reading from sys.env}
}
\value{
boolean field
}
\description{
compares schema fields with dataframe colnames and verifies
}
\examples{
\dontrun{
schema_id <- "assaysch_nIw4yAq8"
schema_type <- "assay-result",
df <- data.frame(
 "plate" = c('davut'),
 "analytes" = c('0.2'),
 'File1' = c('aaaa'),
 check.names = FALSE
}
}
