% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warehouse_disconnect.R
\name{warehouse_disconnect}
\alias{warehouse_disconnect}
\title{Disconnect from a Benchling data warehouse}
\usage{
warehouse_disconnect(conn)
}
\arguments{
\item{conn}{A warehouse connection object of class PqConnection usually 
returned by warehouse_connect()}
}
\value{
TRUE, invisibly
}
\description{
Close an already opened connection to the data warehouse.
}
\examples{
\dontrun{
conn <- warehouse_connect(tenant = "hemoshear-dev", 
    username = Sys.getenv("BENCHLING_WAREHOUSE_USERNAME"),
    password = Sys.getenv("BENCHLING_WAREHOUSE_PASSWORD"))
warehouse_disconnect(conn)    
}
}
