% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bl_get_by_id.R
\name{bl_get_by_id}
\alias{bl_get_by_id}
\title{Generic method for getting Benchling entities by identifier.}
\usage{
bl_get_by_id(bl_class, id = NULL, connector = NULL)
}
\arguments{
\item{bl_class}{Benchling "class".}

\item{id}{Benchling identifier for the object.}

\item{connector}{Benchling API client object or data warehouse
connection. See `bl_api_auth` and `bl_warehouse_connect`.
If a API client is passed, then the function will attempt to 
retrieve the data from the API. If database connection is passed,
then the function will attempt to retrieve the data from the warehouse.}
}
\description{
Benchling identifiers will typically start with three letters, 
followed by an underscore, and end with a series of alphanumeric 
characters. Custom entities will always start with `bfi_`, while other 
"built-in Benchling types" will start with a specific set of three letters.
Some examples of built-in Benchling types include plates, (`plt`), 
entries (`etr`), and containers (`con`).
}
