% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bl_get_by_name.R
\name{bl_get_by_name}
\alias{bl_get_by_name}
\title{Generic method for getting Benchling entities by name.}
\usage{
bl_get_by_name(bl_class, name = NULL, connector = NULL)
}
\arguments{
\item{bl_class}{Benchling "class".}

\item{name}{Name of the Benchling entity to be retrieved.}

\item{connector}{Benchling API client object or data warehouse
connection. See `bl_api_auth` and `bl_warehouse_connect`.
If a API client is passed, then the function will attempt to 
retrieve the data from the API. If database connection is passed,
then the function will attempt to retrieve the data from the warehouse.}
}
\description{
The naming convention of a Benchling entity depend on the naming options 
given to Benchling users in the schema definition. The name might be the 
same as the registry ID, however, it is also possible to allow Benchling 
users to name entities in a free-form manner, or "name templates" can be 
created to allow for entity names to be comprised of fields in the entity.
}
