% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_get_entity.R
\name{bulk_get_entity}
\alias{bulk_get_entity}
\title{Call the bulk-get API endpoints or single API endpoint if no set of bulk-get 
API endpoints is provided for an arbitrary set of entity identifiers.}
\usage{
bulk_get_entity(entity_id)
}
\arguments{
\item{entity_id}{Either a character vector or list that contains the entities. All entities 
in the list or character vector must be character strings.}
}
\value{
A named list, where the names are the identifiers or elements from entity_id and the values
are the responses from the calls made to each identifier's set of bulk-get API endpoints or single
API endpoint in the form of character strings.
}
\description{
bulk_get_entity.R will take a character vector or list of character strings and for each element 
send a request to that element's set of bulk-get API endpoints or single API endpoint if the 
bulk-get API endpoints are written as NA and return a named list containing the responses for 
each request in the form of character strings.
}
\examples{
\dontrun{
entity_id <- c("seq_Cuf0omCm", "bfi_Q1PMlXkf", "loc_43OIpas")
entities <- bulk_get_entity(entity_id)
}
}
