% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundle_api_results.R
\name{.bundle_api_results}
\alias{.bundle_api_results}
\title{Generate a named list of tibble-formatted data tables using the response contents from Benchling
for specific entity identifiers.}
\usage{
.bundle_api_results(entity_responses)
}
\arguments{
\item{entity_responses}{A named list where the names are entity identifiers and each element is a 
list that contains the response contents from the call made to that identifier's Single-Get API 
endpoint in Benchling.}
}
\value{
A named list where the names are entity types and each element is another named list
where the names are entity schema names and each element is a tibble-formatted data table 
displaying the Single-Get API endpoint response contents data extracted Benchling for 
each entity identifier corresponding to a specific entity schema name and entity type.
}
\description{
.bundle_api_results.R takes a named list where the names are entity identifiers and each element
is a list containing that identifier's Single-Get API endpoint response contents from Benchling 
and sorts the data into a named list where the names are the collective entity types for all of 
the entity identifiers in the input and each element is also a named list where the names are 
entity schema names and the elements are tibble-formatted data tables containing the Single-Get 
API endpoint response contents data for each entity identifier associated with a specific schema 
name and entity type.
}
\examples{
\dontrun{
entity_responses <- get_entity(entity_id = c("con_m1dmbdV8", "bfi_Ur5DfvGJ", "seq_Gju61mCm", 
"bfi_Q13AlXkf", "bfi_Ks908uWV", "ent_Ec76qX9f","ent_sPrxBNOh", "box_K9950IQ8", "box_923aklum", 
"bfi_smi97554l"), api_key=Sys.getenv("BENCHLING_API_KEY"))
res <- .bundle_api_results(entity_responses=entity_responses)
}
}
\keyword{internal}
