% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_entity.R
\name{.craft_api_queries}
\alias{.craft_api_queries}
\title{Format each element using its entity identifier's name to generate proper Single-Get API endpoint URLs for 
that identifier}
\usage{
.craft_api_queries(entity_single_get_endpoints)
}
\arguments{
\item{entity_single_get_endpoints}{A named character vector where the names are entity identifiers and the 
elements are the Single-Get API endpoint URL formats for each identifier.}
}
\value{
A named character vector where the names are entity identifiers and the elements are the 
respective Single-Get API endpoint URLs for each identifier.
}
\description{
.craft_api_queries.R takes a named character vector with one or more elements where the names are entity 
identifiers and the elements are each identifier's respective Single-Get API endpoint URL formats and 
rewrites the element using the name of the identifier to generate proper Single-Get API endpoint URLs.
}
\examples{
\dontrun{
entity_single_get_endpoints <- c(
"seq_Cuf0bmCm"="https://hemoshear-dev.benchling.com/api/v2/dna-sequences/ENTITY_ID", 
"bfi_Q1PMlXkf"="https://hemoshear-dev.benchling.com/api/v2/custom-entities/ENTITY_ID", 
"box_7YutniM0"="https://hemoshear-dev.benchling.com/api/v2/boxes/ENTITY_ID",
"bfi_Ks908uWV"="https://hemoshear-dev.benchling.com/api/v2/custom-entities/ENTITY_ID")
res <- .craft_api_queries(entity_single_get_endpoints=entity_single_get_endpoints)
}
}
\keyword{internal}
