% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_entity_data_by_id.R
\name{.get_api_endpoints}
\alias{.get_api_endpoints}
\title{Match each element to its corresponding Single-Get API endpoint URL format}
\usage{
.get_api_endpoints(entity_id, entity_list = NULL)
}
\arguments{
\item{entity_id}{A named character vector where the names are entity identifiers and the elements 
are the entity types associated with those identifiers.}

\item{entity_list}{A list of vectors where each vector designates the schema type and API options for 
the entities we are interested in and the names are the first characters seen in the identifiers for
each entity. If NULL (the default), the function will use a default list.

Alternatively, a custom or more specified list can be used based on the purpose and intent of the list
and overall application of the function.}
}
\value{
A named character vector where the names are entity identifiers and the elements are the
Single-Get API endpoint URL formats associated with them based on the respective entities each identifier 
correlates to and matches with.
}
\description{
.get_api_endpoints.R takes a named character vector with one or more elements where the names are entity 
identifiers and the elements are the entity types for those identifiers and matches each element 
to their respective Single-Get API endpoint URL formats according to which entity the schema correlates 
with.
}
\examples{
\dontrun{
entity_id1 <- c("seq_Cuf0bmCm"="dna_sequence", "bfi_Q1PMlXkf"="custom_entity", "box_7YutniM0"="box")
res1 <- .get_api_endpoints(entity_id=entity_id1, entity_list=NULL)

entity_id2 <- c("bfi_Ur5DfvGJ"="custom_entity", "seq_Gju61mCm"="dna_sequence", 
"bfi_Q13AlXkf"="custom_entity", "bfi_Ks908uWV"="custom_entity", "ent_Ec76qX9f"="user", 
"ent_sPrxBNOh"="user", "box_K9950IQ8"="box")
res2 <- .get_api_endpoints(entity_id=entity_id2, entity_list=NULL)
}
}
\keyword{internal}
