% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_api_responses.R
\name{.handle_api_responses}
\alias{.handle_api_responses}
\title{Check each element's API response and remove any elements that have errors
in their responses.}
\usage{
.handle_api_responses(entity_id_responses)
}
\arguments{
\item{entity_id_responses}{A named list where the names are entity identifiers and
the elements are each entity identifier's Single-Get API endpoint response contents 
from Benchling.}
}
\value{
A named list where the names are valid entity identifiers that do not 
possess errors in their Single-Get API endpoint response contents and the
elements are said identifier's Single-Get API endpoint response contents.
}
\description{
.handle_api_response.R takes a named list where the names are entity identifiers 
and the elements are each identifier's Single-Get API endpoint response contents 
extracted from Benchling and analyzes it to see if any entity identifiers contain 
errors in their Single-Get API endpoint responses and removes them from the list.
}
\examples{
\dontrun{
entity_id <- c("bfi_Ur5DfvGJ", "seq_wPR42xCm", "bfi_Q13AlXkf", "bfi_Ks908uWV", "ent_Ec76qX9f", 
"ent_sPrxBNOh", "wpa_Uj99m230", "cdf_9ppojm5e", "bfi_QueRtY99")
entity_id_responses <- get_entity_data_by_id(entity_id=entity_id, 
api_key=Sys.getenv("BENCHLING_API_KEY"))
res <- .handle_api_responses(entity_id_responses=entity_id_responses)
}
}
