% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_results.R
\name{.validate_blob_link_column_values}
\alias{.validate_blob_link_column_values}
\title{Verify that values in a blob link column refer to valid file paths on the 
local machine.}
\usage{
.validate_blob_link_column_values(
  errors,
  values,
  column_name,
  multi_select_map
)
}
\arguments{
\item{errors}{Errors}

\item{values}{Values in the column.}

\item{column_name}{Name of the column in the input data frame.}
}
\description{
Verify that values in a blob link column refer to valid file paths on the 
local machine.
}
\keyword{internal}
