% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_results.R
\name{.validate_column_values}
\alias{.validate_column_values}
\title{Validate the values in the column of a results table.}
\usage{
.validate_column_values(errors, values, column_name)
}
\description{
If the column refers to an entity link field, the entity IDs in the input
data frame are checked against those currently in the registry. If the column
is a dropdown, then the values are checked against the dropdown options. 
If the column is a storage link, the function checks to see if the IDs in the 
input data frame exist in the inventory. If the column refers to a blob
link field, then the function will assume the values are file paths and will
check to see if the file exists on the local machine.
}
