% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_results.R
\name{.validate_entity_column_values}
\alias{.validate_entity_column_values}
\title{Verify that the identifiers (or names) for an entity link field match
identifiers (or names) that currently exist in the registry.}
\usage{
.validate_entity_column_values(
  conn,
  errors,
  values,
  column_name,
  target_schema_id,
  id_or_name
)
}
\description{
Verify that the identifiers (or names) for an entity link field match
identifiers (or names) that currently exist in the registry.
}
\keyword{internal}
