% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropdown.R
\name{get_dropdown}
\alias{get_dropdown}
\title{Get metadata for a dropdown menu}
\usage{
get_dropdown(conn, name)
}
\arguments{
\item{conn}{Database connection opened with `connect_warehouse`.}

\item{name}{Schema name for the dropdown. This value is stored in the `name`
column of the `dropdown` table in the Benchling data warehouse. 
The name of a dropdown is set when it is created, and it doesn't have to follow
a specific pattern.}
}
\description{
Get metadata for a dropdown menu
}
\examples{
\dontrun{
# Connect to the Benchling data warehouse
conn <- connect_warehouse("hemoshear-dev", 
    username = Sys.getenv("BENCHLING_DEV_WAREHOUSE_USERNAME"),
    password = Sys.getenv("BENCHLING_DEV_WAREHOUSE_PASSWORD"))
# To see all dropdown menus, you can read the entire `dropdown` table. 
# This won't pull the dropdown options that correspond to each dropdown,
# so it probably won't be that large. 
DBI::dbReadTable(conn, "dropdown")
}
}
