% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropdown.R
\name{get_dropdown_options}
\alias{get_dropdown_options}
\title{Get the options in a dropdown menu}
\usage{
get_dropdown_options(conn, schema_id)
}
\arguments{
\item{conn}{Database connection opened with `warehouse_connect`.}

\item{schema_id}{Schema ID for the dropdown.}
}
\value{
character vector of dropdown menu options
}
\description{
Get the options in a dropdown menu
}
