% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_entity.R
\name{get_entity}
\alias{get_entity}
\title{Calls each element's Single-Get API endpoints in Benchling and extracts the response.}
\usage{
get_entity(entity_id, api_key = Sys.getenv("BENCHLING_API_KEY"))
}
\arguments{
\item{entity_id}{A character vector with 1 or more elements.}

\item{api_key}{Benchling API key.}
}
\value{
A named list where the names are entity identifiers and each element is a list that contains
the response contents from the call made to that identifier's Single-Get API endpoint in Benchling.
}
\description{
get_entity.R takes a character vector with one or more elements and for each element calls its Single-Get API 
endpoints in Benchling and extracts the response contents and stores it in a list.
}
\examples{
\dontrun{
entity_id <- c("bfi_Ur5DfvGJ", "seq_Gju61mCm", "bfi_Q13AlXkf", "bfi_Ks908uWV", "ent_Ec76qX9f", 
"ent_sPrxBNOh", "box_K9950IQ8", "box_923aklum")
res <- get_entity(entity_id=entity_id, api_key=Sys.getenv("BENCHLING_API_KEY"))
}
}
