% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plate_diagrams.R
\name{get_plate_diagrams}
\alias{get_plate_diagrams}
\title{Find and parse plate diagrams in a notebook entry}
\usage{
get_plate_diagrams(entry, plate_dim = NULL)
}
\arguments{
\item{entry}{response from GET /entry/{entry_id}}

\item{plate_dim}{A numeric vector of length 1 that describes the number of wells
in the plates that should be read. The plate sizes must be one of the following:
6, 12, 24, 48, 96, 384, and 1536. By default, `plate_dim` is `NULL` and the 
function will return plates of all sizes supported by Benchling.}
}
\value{
List of data frames representing the plate diagrams present
in the notebook entry.
}
\description{
Benchling users can add special unstructured tables to notebook entrys
called "plate diagrams" for conveying information about the contents
of physical plates. The `read_plate_layouts` function will extract these 
plate diagrams into data frames for further processing in R.
}
\examples{
\dontrun{
client <- connect_sdk(tenant="https://hemoshear-dev.benchling.com",
                            api_key=Sys.getenv("BENCHLING_DEV_API_KEY"))
entry <- client$entries$get_entry_by_id("etr_f1bpDIes")
plate_diagrams <- get_plate_diagrams(entry)
}
}
