% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_entity_type.R
\name{infer_entity_type}
\alias{infer_entity_type}
\title{Infer the entity types of elements contained in a vector or list}
\usage{
infer_entity_type(entity_id)
}
\arguments{
\item{entity_id}{Either a character vector or list that contains the entities that we 
are trying to find the schema types for. All entities in the list or character vector 
must be character strings.}
}
\value{
A named character vector, where the names are the identifiers or elements from entity_id
and the values are the schema types.
}
\description{
infer_entity_type.R takes either a character vector or list of character strings and 
for each element tries to infer the schema type using string-matching techniques.
}
\examples{
\dontrun{
entity_id <- c("seq_Cuf0bmCm", "bfi_Q1PMlXkf")
res <- infer_entity_type(entity_id)
}
}
