% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_multiselect_column.R
\name{list_multiselect_columns}
\alias{list_multiselect_columns}
\title{List the multi-select columns in a data frame retrieved from the data 
warehouse}
\usage{
list_multiselect_columns(conn, df)
}
\arguments{
\item{conn}{Database connection opened with `warehouse_connect`.}

\item{df}{Data frame retrieved from the data warehouse.}
}
\value{
Character vector where the names are the names of the multi-select
columns in the data frame and the values are the positions of the multi-select
columns in the data frame.
}
\description{
List the multi-select columns in a data frame retrieved from the data 
warehouse
}
\examples{
\dontrun{
conn <- warehouse_connect("hemoshear-dev", 
    username = Sys.getenv("BENCHLING_DEV_WAREHOUSE_USERNAME"),
    password = Sys.getenv("BENCHLING_DEV_WAREHOUSE_PASSWORD"))
res <- DBI::dbGetQuery(conn, "SELECT * FROM simple_plate_analyte_mapping$raw")
list_multiselect_columns(conn, res)
}
}
