% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_entry_tables.R
\name{read_entry_table}
\alias{read_entry_table}
\title{Read unstructured table in a notebook entry}
\usage{
read_entry_table(entry, day, table_index)
}
\arguments{
\item{entry}{Notebook entry in JSON format.}

\item{day}{Integer for the day in the notebook entry. See `find_entry_tables`.}

\item{table_index}{Integer for the position of the table in the notebook entry list. 
See `find_entry_tables`.}
}
\value{
Data frame representing the unstructured table in the notebook
entry.
}
\description{
Read unstructured table in a notebook entry
}
\examples{
\dontrun{
entry <- get_entry(id = "etr")
table_indices <- find_entry_tables(entry)
# This code will read the first table found in the first day of the notebook
# entry.
a_table <- read_table(entry, day=1, table_index=table_indices[[1]][1])
}
}
