% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_entry_tables.R
\name{read_entry_tables}
\alias{read_entry_tables}
\title{Read all unstructured tables in a notebook entry.}
\usage{
read_entry_tables(
  entry,
  day = NULL,
  table_position = NULL,
  table_name = NULL,
  return_table_name = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{entry}{Notebook entry in JSON format. See `get_entry`.}

\item{day}{Integer for the day in the notebook entry. See `find_entry_tables`.}

\item{table_position}{Integer for the position of the table in the notebook entry list. 
See `find_entry_tables`.}

\item{table_name}{A specific table can be read by name with the `table_name` argument.
If `table_name` is provided, then `day` and `table_position` are ignored.}

\item{return_table_name}{If return_return_table_name is TRUE, then the names of the tables
in the notebook entry will be returned as names in the output list.}

\item{verbose}{If verbose is TRUE, then the function will alert the user
if no tables can be found for some days in the notebook.}
}
\value{
List of data frames representing the unstructured tables in
the notebook entry.
}
\description{
This function will read all unstructured tables in a notebook entry,
returning a list of data frames.
}
\examples{
\dontrun{
client <- benchlingr::benchling_api_auth(
    tenant="https://hemoshear-dev.benchling.com",
    api_key = Sys.getenv("BENCHLING_DEV_API_KEY"))
entry <- client$entries$get_entry_by_id("etr_IWLMFYhR")

entry <- client$entries$get_entry_by_id("etr_T3WZTyAe")
tables <- read_entry_tables(entry, table_name='Experimental Conditions')
entry <- client$entries$get_entry_by_id("etr_aTa6eDWA")
tables <- read_entry_tables(entry, table_name=c("TableA", "TableB"))
}
}
