% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_entry_tables.R
\name{read_entry_tables}
\alias{read_entry_tables}
\title{Read all unstructured tables in a notebook entry.}
\usage{
read_entry_tables(entry, verbose = FALSE)
}
\arguments{
\item{entry}{Notebook entry in JSON format. See `get_entry`.}

\item{verbose}{If verbose, then the function will alert the user
if no tables can be found for some days in the notebook.}
}
\value{
List of data frames representing the unstructured tables in
the notebook entry.
}
\description{
This function will read all unstructured tables in a notebook entry,
returning a list of data frames.
}
\examples{
\dontrun{
client <- benchling_api_auth(tenant="https://hemoshear.benchling.com")
client$entries$get_entry_by_id
entry <- get_entry(id = "etr")
tables <- read_tables(entry)
}
}
