% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_plate_diagrams.R
\name{read_plate_diagrams}
\alias{read_plate_diagrams}
\title{Find and parse plate diagrams in a notebook entry}
\usage{
read_plate_diagrams(entry)
}
\arguments{
\item{entry}{response from GET /entry/{entry_id}}
}
\value{
List of data frames representing the plate diagrams present
in the notebook entry.
}
\description{
Benchling users can add special unstructured tables to notebook entrys
called "plate diagrams" for conveying information about the contents
of physical plates. The `read_plate_layouts` function will extract these 
plate diagrams into data frames for further processing in R.
}
